Add getBlockVisibilityLabel utility function and refactor visibility label logic in list view components
- Introduced `getBlockVisibilityLabel` function to generate human-readable labels for block visibility based on viewport settings.
- Refactored visibility label determination in `ListViewBlockSelectButton` and `ListViewBlock` components to utilize the new utility function, improving code clarity and maintainability.
