+++ packages/block-editor/src/components/block-visibility/index.js
@@ -3,3 +3,4 @@ export { default as useBlockVisibility } from './use-block-visibility';
 export { default as ViewportVisibilityToolbar } from './viewport-toolbar';
 export { default as BlockVisibilityViewportMenuItem } from './viewport-menu-item';
 export { default as ViewportVisibilityInfo } from './viewport-visibility-info';
+export { getBlockVisibilityLabel } from './utils';
